+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphand/cli
 
+## 1.2.9
+
+### Patch Changes
+
+- improve ValidationError class
+- Updated dependencies
+  - @graphand/client@1.2.13
+
 ## 1.2.8
 
 ### Patch Changes